+++ packages/@expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
@@ -10,8 +10,11 @@ import { ResolutionContext } from 'metro-resolver';
 import path from 'path';
 import resolveFrom from 'resolve-from';
 
+import { env } from '../../../utils/env';
+import { WebSupportProjectPrerequisite } from '../../doctor/web/WebSupportProjectPrerequisite';
 import { PlatformBundlers } from '../platformBundlers';
 import { importMetroResolverFromProject } from './resolveFromProject';
+import { getAppRouterRelativeEntryPath } from './router';
 
 function withWebPolyfills(config: ConfigT): ConfigT {
   const originalGetPolyfills = config.serializer.getPolyfills
@@ -102,21 +105,43 @@ export function withWebResolvers(config: ConfigT, projectRoot: string) {
     },
   };
 
+  const preferredMainFields: { [key: string]: string[] } = {
+    // Defaults from Expo Webpack. Most packages using `react-native` don't support web
+    // in the `react-native` field, so we should prefer the `browser` field.
+    // https://github.com/expo/router/issues/37
+    web: ['browser', 'module', 'main'],
+  };
+
   return withCustomResolvers(config, projectRoot, [
     // Add a resolver to alias the web asset resolver.
     (immutableContext: ResolutionContext, moduleName: string, platform: string | null) => {
-      const context = { ...immutableContext };
+      const context = { ...immutableContext } as ResolutionContext & { mainFields: string[] };
 
       // Conditionally remap `react-native` to `react-native-web`
       if (platform && platform in extraNodeModules) {
         context.extraNodeModules = extraNodeModules[platform];
       }
 
+      const mainFields = env.EXPO_METRO_NO_MAIN_FIELD_OVERRIDE
+        ? context.mainFields
+        : platform && platform in preferredMainFields
+        ? preferredMainFields[platform]
+        : context.mainFields;
+
       const result = resolve(
         {
           ...context,
           preferNativePlatform: platform !== 'web',
           resolveRequest: undefined,
+
+          // Passing `mainFields` directly won't be considered
+          // we need to extend the `getPackageMainPath` directly to
+          // use platform specific `mainFields`.
+          getPackageMainPath(packageJsonPath) {
+            // @ts-expect-error: mainFields is not on type
+            const package_ = context.moduleCache.getPackage(packageJsonPath);
+            return package_.getMain(mainFields);
+          },
         },
         moduleName,
         platform
@@ -140,16 +165,30 @@ export function withWebResolvers(config: ConfigT, projectRoot: string) {
 }
 
 /** Add support for `react-native-web` and the Web platform. */
-export function withMetroMultiPlatform(
+export async function withMetroMultiPlatformAsync(
   projectRoot: string,
   config: ConfigT,
   platformBundlers: PlatformBundlers
 ) {
-  // Bail out early for performance enhancements if web is not enabled.
-  if (platformBundlers.web !== 'metro') {
+  // Auto pick App entry: this is injected with Babel.
+  process.env.EXPO_ROUTER_APP_ROOT = getAppRouterRelativeEntryPath(projectRoot);
+  process.env.EXPO_PROJECT_ROOT = process.env.EXPO_PROJECT_ROOT ?? projectRoot;
+
+  if (platformBundlers.web === 'metro') {
+    await new WebSupportProjectPrerequisite(projectRoot).assertAsync();
+  } else {
+    // Bail out early for performance enhancements if web is not enabled.
     return config;
   }
 
+  return withMetroMultiPlatform(projectRoot, config, platformBundlers);
+}
+
+function withMetroMultiPlatform(
+  projectRoot: string,
+  config: ConfigT,
+  platformBundlers: PlatformBundlers
+) {
   let expoConfigPlatforms = Object.entries(platformBundlers)
     .filter(([, bundler]) => bundler === 'metro')
     .map(([platform]) => platform);