+++ packages/expo-updates/ios/EXUpdates/ReactDelegateHandler/ExpoUpdatesReactDelegateHandler.swift
@@ -2,6 +2,13 @@
 
 import ExpoModulesCore
 
+/**
+ * Manages and controls the auto-setup behavior of expo-updates in applicable environments.
+ *
+ * In order to deal with the asynchronous nature of updates startup, this class creates dummy
+ * RCTBridge and RCTRootView objects to return to the ReactDelegate, replacing them with the real
+ * objects when expo-updates is ready.
+ */
 public class ExpoUpdatesReactDelegateHandler: ExpoReactDelegateHandler, EXUpdatesAppControllerDelegate, RCTBridgeDelegate {
   private weak var reactDelegate: ExpoReactDelegate?
   private var bridgeDelegate: RCTBridgeDelegate?